requirements.txt
fastapi
uvicorn
python-multipart
supabase
pydantic
python-dotenv
google-genai